requirements.txt
python-decouple
uptime
pusher